requirements.txt
Basilica==0.2.8
Blueprint==3.4.2
Flask==1.1.1
Flask_Migrate==2.5.2
Flask_SQLAlchemy==2.4.1
Jinja2==2.11.1
jsonify==0.5
numpy==1.18.1
python-dotenv==0.11.0
request==2019.4.13
requests==2.23.0
scikit-learn==0.22.1
gunicorn==20.0.4
tweepy==3.8.0
psycopg2==2.8.4